GIFT checklists

Retrieve GIFT checklists that fulfill specific criteria.

Usage

GIFT_checklist_conditional(
  taxon_name = "Tracheophyta",
  ref_included = c("all", "native", "native and naturalized",
    "native and historically introduced", "endangered", "endemic", "naturalized",
    "other subset")[1:4],
  type_ref = c("Account", "Catalogue", "Checklist", "Flora", "Herbarium collection",
    "Key", "Red list", "Report", "Species Database", "Survey"),
  entity_class = c("Island", "Island/Mainland", "Mainland", "Island Group",
    "Island Part"),
  native_indicated = FALSE,
  natural_indicated = FALSE,
  end_ref = FALSE,
  end_list = FALSE,
  suit_geo = FALSE,
  complete_taxon = TRUE,
  GIFT_version = "latest",
  api = "http://gift.uni-goettingen.de/api/extended/",
  list_set = NULL,
  taxonomy = NULL
)

Arguments

taxon_name

Character string corresponding to the taxonomic group of interest.

ref_included

Character, options are 'all', 'native', 'native and naturalized', 'native and historically introduced', 'endangered', 'endemic', 'naturalized', 'other subset'

type_ref

Character, options are 'Account', 'Catalogue', 'Checklist', 'Flora', 'Herbarium collection', 'Key', 'Red list', 'Report', 'Species Database', 'Survey'

entity_class

Character, options are 'Island', 'Island/Mainland', 'Mainland', 'Island Group', 'Island Part'

native_indicated

Boolean, whether only lists where native status is available should be retrieved.

natural_indicated

Boolean, whether only lists where natural status is available should be retrieved.

end_ref

Boolean, whether only lists where endemism at the reference level is available should be retrieved.

end_list

Boolean, whether only lists where endemism at the list level is available should be retrieved.

suit_geo

Boolean, whether only suitable polygons should be retrieved.

complete_taxon

Boolean, default TRUE.

GIFT_version

character string defining the version of the GIFT database to use. The function retrieves by default the most up-to-date version.

api

character string defining from which API the data will be retrieved.

list_set

default NULL. If not, a vector of identification numbers of GIFT checklists.

taxonomy

default NULL. If not, it has to be the taxonomy table (see GIFT_taxonomy()).

Value

A data frame with 16 columns.

Details

Here is what each column refers to: 'ref_ID' - Identification number of each reference. 'type'- What type the source is. 'subset'- What information regarding the status of species is available. 'native_indicated'- Whether native status of species is available in the source. 'natural_indicated' - Whether naturalized status of species is available in the source. 'end_ref' - Whether endemism information is available in the source. 'restricted' - Whether the access to this reference is restricted. 'taxon_ID'- Identification number of species. 'list_ID'- Identification number of each list. 'end_list' - Whether endemism information is available in the list. 'entity_ID'- Identification number of the polygon of the list. 'geo_entity'- Name of the location. 'suit_geo'- Is the polygon suitable. 'entity_class'- Type of polygon. 'entity_type'- Name of the location. 'taxon_name'- Name of the group of taxa available.

References

Weigelt, P, König, C, Kreft, H. GIFT – A Global Inventory of Floras and Traits for macroecology and biogeography. J Biogeogr. 2020; 47: 16– 43. https://doi.org/10.1111/jbi.13623

See also

Examples

if (FALSE) {
ex <- GIFT_checklist_conditional(taxon_name = "Embryophyta", 
ref_included = c("all", "native", "native and naturalized",
"native and historically introduced", "endangered",
"endemic", "naturalized", "other subset")[1:4],
type_ref = c("Account", "Catalogue", "Checklist","Flora",
"Herbarium collection", "Key", "Red list", "Report", "Species Database",
 "Survey"),
 entity_class = c("Island", "Island/Mainland", "Mainland", "Island Group",
 "Island Part"),
 native_indicated = FALSE, natural_indicated = FALSE, end_ref = FALSE,
 end_list = FALSE, suit_geo = TRUE, complete_taxon = TRUE,
 list_set = NULL, taxonomy = NULL)

}