Taxonomy of GIFT

Retrieves the taxonomy of GIFT.

Usage

GIFT_taxonomy(
  GIFT_version = "latest",
  api = "http://gift.uni-goettingen.de/api/extended/"
)

Arguments

GIFT_version

character string defining the version of the GIFT database to use. The function retrieves by default the most up-to-date version.

api

character string defining from which API the data will be retrieved.

Value

A data frame with 6 columns.

Details

Column taxon_ID indicates the identification number of each taxonomic entry. The names describing taxa are in the column taxon_name.The third column spells the author name for a given taxon. The column taxon_lvl splits every taxon in genus, family, order or superior orders. Taxonomy is a linear sequence of left and right borders for each taxon. This is nested, for example left and right borders of a genus would fall between the left and right borders of the corresponding family. Columns lft and rgt respectively refer to these left and right borders.

References

 Weigelt, P, König, C, Kreft, H. GIFT – A Global Inventory of Floras and
 Traits for macroecology and biogeography. J Biogeogr. 2020; 47: 16– 43.
 https://doi.org/10.1111/jbi.13623

See also

Examples

if (FALSE) {
ex <- GIFT_taxonomy()
}